// src/types/socket.ts
export interface SocketData {
  userId: string;
  username: string;
  currentGameId?: string;
  spectatingGameId?: string;
  isInQueue: boolean;
  matchInterval?: NodeJS.Timeout;
}